Refactored based on PR feedback: removed SceneController child GO, Added WorldBlockersController instantiation with DI in SceneController initialization, Added a test and unified their setup, removed old green blockers code and transformed BlockerHandler to the new world blockers logic
