--- unity-client/Assets/Scripts/MainScripts/DCL/Controllers/Scene/Tests/WorldBlockerHandlerShould.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5edb633a52a1e41a4a6b51b471555652
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
