--- unity-client/Assets/Scripts/MainScripts/DCL/Controllers/Scene/WorldBlockerHandler.cs
@@ -1,86 +1,0 @@
-﻿using DCL.Configuration;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace DCL.Controllers
-{
-    public class WorldBlockerHandler : BlockerHandler
-    {
-        HashSet<Vector2Int> blockersToRemove = new HashSet<Vector2Int>();
-        HashSet<Vector2Int> blockersToAdd = new HashSet<Vector2Int>();
-
-        public void SetupGlobalBlockers(HashSet<Vector2Int> allLoadedParcelCoords, float height, Transform parent)
-        {
-            if (allLoadedParcelCoords.Count == 0) return;
-
-            blockersToRemove.Clear();
-            blockersToAdd.Clear();
-
-            auxScaleVec.x = ParcelSettings.PARCEL_SIZE;
-            auxScaleVec.y = height;
-            auxScaleVec.z = ParcelSettings.PARCEL_SIZE;
-
-            auxPosVec.y = (height - 1) / 2;
-
-            // Detect blockers to be removed
-            foreach (var item in blockers)
-            {
-                if (allLoadedParcelCoords.Contains(item.Key))
-                {
-                    blockersToRemove.Add(item.Key);
-                }
-                else
-                {
-                    bool foundAroundLoadedScenes = false;
-                    for (int i = 0; i < aroundOffsets.Length; i++)
-                    {
-                        Vector2Int offset = aroundOffsets[i];
-                        Vector2Int checkedPosition = new Vector2Int(item.Key.x + offset.x, item.Key.y + offset.y);
-
-                        if (allLoadedParcelCoords.Contains(checkedPosition))
-                        {
-                            foundAroundLoadedScenes = true;
-                            break;
-                        }
-                    }
-
-                    if (!foundAroundLoadedScenes)
-                        blockersToRemove.Add(item.Key);
-                }
-            }
-
-            // Detect missing blockers to be added
-            using (var it = allLoadedParcelCoords.GetEnumerator())
-            {
-                while (it.MoveNext())
-                {
-                    Vector2Int pos = it.Current;
-
-                    for (int i = 0; i < aroundOffsets.Length; i++)
-                    {
-                        Vector2Int offset = aroundOffsets[i];
-                        Vector2Int checkedPosition = new Vector2Int(pos.x + offset.x, pos.y + offset.y);
-
-                        if (!allLoadedParcelCoords.Contains(checkedPosition) && !blockers.ContainsKey(checkedPosition))
-                        {
-                            blockersToAdd.Add(checkedPosition);
-                        }
-                    }
-                }
-            }
-
-            // Remove extra blockers
-            foreach (var coords in blockersToRemove)
-            {
-                blockers[coords].Release();
-                blockers.Remove(coords);
-            }
-
-            // Add missing blockers
-            foreach (var coords in blockersToAdd)
-            {
-                InstantiateBlocker(coords, parent);
-            }
-        }
-    }
-}
