--- unity-client/Assets/Scripts/MainScripts/DCL/Controllers/Scene/WorldBlockerHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff09846aa2ee240c18b52ce7019c0cdb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
