+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/WorldBlockersController.cs
@@ -3,16 +3,23 @@ using UnityEngine;
 
 namespace DCL.Controllers
 {
-    public class WorldBlockersController : MonoBehaviour
+    public class WorldBlockersController
     {
-        WorldBlockerHandler blockerHandler;
+        SceneController sceneController;
+        Transform blockersParent;
+        BlockerHandler blockerHandler;
         HashSet<Vector2Int> allLoadedParcelCoords = new HashSet<Vector2Int>();
 
-        void Start()
+        public WorldBlockersController(SceneController sceneController, Transform blockersParent)
         {
-            blockerHandler = new WorldBlockerHandler();
+            blockerHandler = new BlockerHandler();
 
-            SceneController.i.OnNewSceneAdded += OnNewSceneAdded;
+            this.sceneController = sceneController;
+
+            blockersParent.position = Vector3.zero;
+            this.blockersParent = blockersParent;
+
+            sceneController.OnNewSceneAdded += OnNewSceneAdded;
             DCLCharacterController.i.characterPosition.OnPrecisionAdjust += OnWorldReposition;
 
             SceneController.OnDebugModeSet += () =>
@@ -23,7 +30,7 @@ namespace DCL.Controllers
 
         void OnDisable()
         {
-            SceneController.i.OnNewSceneAdded -= OnNewSceneAdded;
+            sceneController.OnNewSceneAdded -= OnNewSceneAdded;
             DCLCharacterController.i.characterPosition.OnPrecisionAdjust -= OnWorldReposition;
         }
 
@@ -32,25 +39,25 @@ namespace DCL.Controllers
             allLoadedParcelCoords.Clear();
 
             // Create fast (hashset) collection of loaded parcels coords
-            foreach (var element in SceneController.i.loadedScenes)
+            foreach (var element in sceneController.loadedScenes)
             {
                 if (!element.Value.isReady) continue;
 
                 allLoadedParcelCoords.UnionWith(element.Value.parcels);
             }
 
-            blockerHandler.SetupGlobalBlockers(allLoadedParcelCoords, 100, transform);
+            blockerHandler.SetupGlobalBlockers(allLoadedParcelCoords, 100, blockersParent);
         }
 
         void OnWorldReposition(DCLCharacterPosition charPos)
         {
-            var newPosition = charPos.WorldToUnityPosition(Vector3.zero);
-            transform.position = newPosition;
+            var newPosition = charPos.WorldToUnityPosition(Vector3.zero); // Blockers parent original position
+            blockersParent.position = newPosition;
         }
 
         void OnNewSceneAdded(ParcelScene newScene)
         {
-            if (SceneController.i.isDebugMode) return;
+            if (sceneController.isDebugMode) return;
 
             newScene.OnSceneReady += OnSceneReady;
         }
@@ -59,7 +66,7 @@ namespace DCL.Controllers
         {
             scene.OnSceneReady -= OnSceneReady;
 
-            if (SceneController.i.isDebugMode)
+            if (sceneController.isDebugMode)
             {
                 blockerHandler.CleanBlockers();
                 return;
